// Interfaces/Interfaces.ts
export interface CommandItemInterface {
  bind: Record<string, string>
  bind_name: Record<string, ValueInterface>
  created_at: string
  currency: string
  id: string
  order_id: string
  price_unit: number
  product_id: string
  quantity: number
  status: string
  store_id: string
  updated_at: string
  product?: ProductInterface,
}
// ...
export interface CategoryInterface {
  id: string,
  name: string,
  description: string,
  parent_category_id?: string | null
  store_id: string,
  slug: string,
  product_count?: number
  view: (string | Blob)[],
  icon: (string | Blob)[],
  created_at: string,
  updated_at: string
  is_visible?: boolean
}

export interface ProductInterface {
  id: string;
  store_id: string;
  categories_id: string[];
  name: string;
  default_feature_id: string;
  slug: string,
  description: string;
  barred_price: number;
  price: number;
  rating: number,
  comment_count: number,
  stock?: number | null
  is_visible: boolean,// TODO
  currency: string;
  created_at: Date;
  updated_at: Date;
  features?: FeatureInterface[]
  categories?: CategoryInterface[]
};


export interface ValueInterface {
  id: string;
  feature_id: string;
  views?: (string | Blob)[] | null;
  icon?: (string | Blob)[] | null;
  text?: string | null;
  key?: string | null;
  stock?: number | null
  additional_price?: number | null
  decreases_stock?: boolean,
  continue_selling?: boolean
  index: number;
  created_at?: string | Date;
  updated_at?: string | Date;
  _request_mode?: 'edited' | 'new'
};

export interface FeatureInterface {
  id: string,
  product_id: string,
  name: string,
  type: string,
  icon?: (string | Blob)[],
  required: boolean,
  regex?: string,
  min?: number,
  max?: number,
  min_size?: number,
  max_size?: number,
  index?: number,
  multiple?: false,
  is_double?: false,
  default?: string | null,
  is_default: boolean,
  created_at: string,
  updated_at: string,
  values?: ValueInterface[];
  _request_mode?: 'edited' | 'new'
};